+++ apps/api/src/comments/comments.module.ts
@@ -1,10 +1,11 @@
 import { Module } from '@nestjs/common';
 import { ActivityModule } from '../activity/activity.module';
+import { NotificationsModule } from '../notifications/notifications.module';
 import { CommentsController } from './comments.controller';
 import { CommentsService } from './comments.service';
 
 @Module({
-  imports: [ActivityModule],
+  imports: [ActivityModule, NotificationsModule],
   controllers: [CommentsController],
   providers: [CommentsService],
   exports: [CommentsService],