+++ apps/api/src/reactions/reactions.module.ts
@@ -1,10 +1,11 @@
 import { Module } from '@nestjs/common';
 import { ActivityModule } from '../activity/activity.module';
+import { NotificationsModule } from '../notifications/notifications.module';
 import { ReactionsController } from './reactions.controller';
 import { ReactionsService } from './reactions.service';
 
 @Module({
-  imports: [ActivityModule],
+  imports: [ActivityModule, NotificationsModule],
   controllers: [ReactionsController],
   providers: [ReactionsService],
   exports: [ReactionsService],